mv Profiler.updated & assoc helpers into (compile-common/utils)
