+++ packages/compile-common/utils/findUpdatedFiles.js
@@ -38,6 +38,4 @@ const findUpdatedFiles = (
     .filter(file => file);
 };
 
-module.exports = {
-  findUpdatedFiles
-};
+module.exports = findUpdatedFiles;
