+++ packages/compile-common/utils/minimumUpdatedTimePerSource.js
@@ -27,6 +27,4 @@ const minimumUpdatedTimePerSource = sourceFilesArtifacts => {
   return sourceFilesArtifactsUpdatedTimes;
 };
 
-module.exports = {
-  minimumUpdatedTimePerSource
-};
+module.exports = minimumUpdatedTimePerSource;
