+++ packages/compile-common/utils/readAndParseArtifactFiles.js
@@ -52,6 +52,4 @@ const readAndParseArtifactFiles = (sourceFiles, contracts_build_directory) => {
   return sourceFilesArtifacts;
 };
 
-module.exports = {
-  readAndParseArtifactFiles
-};
+module.exports = readAndParseArtifactFiles;
